Options
Argparse was imported to make adding options to this script much easier. Now a user can select to decrypt or encrypt their message (a moot point considering how rot13 works but good practice). The script needs to be cleaned up still and a default option will need to be introduced in the event the user does not select a flag. I may also opt to add inline support to supply the message to encrypt rather than sending the user through the script.