Update command line protocol
The latest updates to the program make the run method obsolete. Currently the program works in line from the terminal when passing a string and a flag. The next additions will be to allow the user to pass a file to the program in addition to the option to pass a string.